refactor : 동시성 및 성능에 대한 테스트 코드의 폴더 위치 변경
